Bot/requirements.txt
streamlit
openai
pinecone-client
huggingface-hub